Finished basic watchlist

+++ app/src/main/java/com/company/vesper/watchlist/WatchListItem.java
@@ -1,19 +1,20 @@
 package com.company.vesper.watchlist;
 
-import org.w3c.dom.NameList;
-
+/**
+ * Holder Item for watchlist
+ */
 public class WatchListItem {
     public String Ticker;
     public String Name;
-    public float closingPrice;
-    public float dailyChange;
+    public double closingPrice;
+    public double dailyChange;
 
     public WatchListItem(String Ticker) {
         // Simple constructor for now
         // As we build the app more we can modify for the rest of the parameter
         this.Ticker = Ticker;
     }
-    public WatchListItem(String Ticker, String Name, float closingPrice, float dailyChange )  {
+    public WatchListItem(String Ticker, String Name, double closingPrice, double dailyChange )  {
         // Constructor for complete version
         this.Ticker = Ticker;
         this.Name = Name;
